chore: move form-publish component
