+++ src/components/form-publish/form-publish-configurer.js
@@ -1,5 +1,5 @@
 import { reduxForm } from 'redux-form';
-import { asyncValidateDatasetInvokePatch } from '../formsLib/asyncValidateDatasetInvokePatch';
+import { asyncValidateDatasetInvokePatch } from '../../pages/dataset-registration-page/formsLib/asyncValidateDatasetInvokePatch';
 
 export const formPublishConfigurer = reduxForm({
   form: 'datasetPublish',
